push

+++ app/src/main/java/com/buxingzhe/pedestrian/bean/activity/WalkRecordInfo.java
@@ -55,7 +55,7 @@ public class WalkRecordInfo implements Parcelable {
     private String createTime;
     private UserBaseInfo user;
     private int stepCount;
-    private int distance;
+    private double distance;
     private String duration;
     private String altitudeAsend;
     private String altitudeHigh;
@@ -86,7 +86,7 @@ public class WalkRecordInfo implements Parcelable {
         createTime = in.readString();
         user = in.readParcelable(UserBaseInfo.class.getClassLoader());
         stepCount = in.readInt();
-        distance = in.readInt();
+        distance = in.readDouble();
         duration = in.readString();
         altitudeAsend = in.readString();
         altitudeHigh = in.readString();
@@ -157,11 +157,11 @@ public class WalkRecordInfo implements Parcelable {
         this.stepCount = stepCount;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
 
@@ -368,7 +368,7 @@ public class WalkRecordInfo implements Parcelable {
         parcel.writeString(createTime);
         parcel.writeParcelable(user, i);
         parcel.writeInt(stepCount);
-        parcel.writeInt(distance);
+        parcel.writeDouble(distance);
         parcel.writeString(duration);
         parcel.writeString(altitudeAsend);
         parcel.writeString(altitudeHigh);
